requirements.txt
fastapi 
uvicorn 
#opentelemetry-proto
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
#opentelemetry-instrumentation
#opentelemetry-instrumentation-fastapi
protobuf
mysql.connector
#jinja2 
#aiofiles
pytz
psycopg2-binary
gunicorn